Use Alchemy::Solidus namespace for decorator constants
In a world where decorators are managed with Zeitwerk, we want to make
sure the naming of our decorators does not conflict with the naming of
decorators in other gems.
