+++ app/decorators/models/alchemy/solidus/spree_taxon_decorator.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Alchemy
+  module Solidus
+    module SpreeTaxonDecorator
+      def self.prepended(base)
+        base.include AlchemySolidus::TouchAlchemyIngredients
+        base.has_many :alchemy_ingredients, class_name: "Alchemy::Ingredients::SpreeTaxon", as: :related_object, dependent: :nullify
+      end
+
+      ::Spree::Taxon.prepend self
+    end
+  end
+end
